refactor: move files related to Annotation
