+++ giraffe/src/components/Annotation/AnnotationLine.scss.d.ts
