+++ giraffe/src/components/Annotation/index.ts
@@ -1,0 +1,1 @@
+export {AnnotationLayer} from './AnnotationLayer'
